refactor: localStorage 사용을 추상화하고 user 훅을 하나로 통합
