+++ src/contexts/TokenContext.tsx
@@ -1,9 +1,18 @@
 import { createContext, ReactNode, useContext, useState } from "react";
 
+const tokenStorage = {
+  get: () => {
+    return localStorage.getItem("access_token") || "";
+  },
+  set: (accessToken: string) => {
+    localStorage.setItem("access_token", accessToken);
+  },
+};
+
 interface TokenContextType {
   isLogin: boolean;
   accessToken: string;
-  setAccessToken: React.Dispatch<React.SetStateAction<string>>;
+  setAccessToken: (accessToken: string) => void;
 }
 
 const TokenContext = createContext<TokenContextType>({
@@ -13,12 +22,16 @@ const TokenContext = createContext<TokenContextType>({
 });
 
 export const TokenContextProvider = ({ children }: { children: ReactNode }) => {
-  const [accessToken, setAccessToken] = useState<string>(localStorage.getItem("access_token") || "");
+  const [accessToken, setAccessToken] = useState<string>(tokenStorage.get());
 
   const value = {
     isLogin: accessToken !== "",
     accessToken,
-    setAccessToken,
+    setAccessToken: (accessToken: string) => {
+      tokenStorage.set(accessToken);
+
+      setAccessToken(accessToken);
+    },
   };
 
   return <TokenContext.Provider value={value}>{children}</TokenContext.Provider>;
